remove default repo_dir value
